Set up behave testing
For UELC, I'm trying out the behave-django package instead
of django-behave. The maintainers of behave-django are actively
using it, unlike django-behave, so it's bound to
be a little more responsive to questions like e.g., how to run
a single test:
  https://github.com/django-behave/django-behave/issues/58

Also, django-behave's custom test runner is kind of awkward
because it doesn't clear the database between scenarios, leading
to confusing tests. behave-django claims to address this problem:
  https://github.com/mixxorz/behave-django/wiki/Getting-started#database-transactions-per-scenario

The tests can be run locally with `make behave` and are set
up to run on travis.