Increase submenu width from 130% to 160%
And some syntax cleanup.

I spent a long time trying to get this submenu to actually fill
the width based on its contents, but I couldn't figure it out.
For example, if changing the submenu to `position: relative;`
invokes the correct width, but causes problems with its parent
menu.

Submenus aren't officially supported by Bootstrap as of version 3,
which is why this is kind of messy.