Fix password bug - PMT #102256
This fixes a bug that Paul noticed - the password that
users get created with isn't getting set correctly.

This happened because I changed the user creation code
a few weeks ago to use create_user() instead of create().
This makes the call to make_password() superfluous and
incorrect.

I've added some tests to ensure this view has the correct
behavior.