Address clone form 'name' problem - PMT #104940
The main problem here was in DynamicHierarchyMixin, and I've added more
thorough testing to how the clone form behaves. The problem was that
DynamicHierarchyMixin was treating a slugified version of the base_url
as the hierarchy's name, using that to look it up, which triggered
pagetree to create a new hierarchy if it didn't exist. As a result, the
clone form would create *two* hierarchies, one with all the data, and
one empty one, unless the name and the base_url were both the same,
i.e.: name: 'case-one', base_url: '/pages/base-url/'

The other bullet point in this PMT has already been implemented. You
don't need to type `/pages/` around the base_url, you can just give it a
slug.