uelc.main.helper_functions.get_user_last_location
Adding unit tests and revisiting the logic here.

Including the hierarchy in the filter args to ensure
the resulting UserLocations are applicable.

Pagetree's UserLocation has a unique constraint on `(user, hierarchy)`.
The intent to "fix up" the UserLocation if a path does not translate to
a Section is good, but can be simplified to simply removing the UserLocation
for the hierarchy. (The confusion here might've been UserLocations from
other hierarchies.)