generate simpler windsock tokens
PMT #109157

Using the hierarchy name is bad because it can (and apparently
frequently does) include a ':' which windsock uses as a delimiter for
tokens. The token gets included directly in the template without any
escaping, so it would also break at the JS level if the name of a
hierarchy happened to include a single quote.

This changes it to just use a simpler alphanumeric form.