add windsock connection smoketest
We already had a smoketest that checked that it could send a message to
the broker. This adds another smoketest that actually generates a token
and makes a connection to the websockets server. This way, if windsock
is down, or if something changes in the token generation that makes
tokens invalid (eg, if clocks get out of skew), the smoketests will
fail.

I already did the same thing (and self-merged so I could be sure) for
tala:

https://github.com/ccnmtl/tala/pull/292